Automatically exclude matches if Box with light=1 is hit
In case of duplicates automatically generate 'exclude list' when Box
with light=1 is hit to properly model the knowledge that the duplicate
is not also matched to this pair.

To still be able to model hitting the duplicate pair in a Box, there is
the key `noExclude` to avoid this automatically generating. This way we
can still go the way of adding a (hidden) virtual additional Box with lights=1